+++ src/lang/date.js
@@ -11,5 +11,17 @@ export default {
       hour: 'numeric',
       minute: 'numeric'
     }
+  },
+  fr: {
+    short: {
+      year: '2-digit', month: '2-digit', day: '2-digit'
+    },
+    long: {
+      year: '2-digit',
+      month: '2-digit',
+      day: '2-digit',
+      hour: 'numeric',
+      minute: 'numeric'
+    }
   }
 }
